requirements.txt
fastapi
uvicorn
python-dotenv
requests
pytz
google-generativeai
duckduckgo-search
